final rollback for weightsensor

+++ Components/itemView.py
@@ -64,49 +64,52 @@ class Ui_MainWindowItemView(object):
                     self.removeProduct(reference_number, barcode)
                     os.remove("predicted_class.txt")
                     self.resumeScanningMessage()
-
     def paused(self):
         if not self.weight_sensor.verify:
-            self.resumeScanningMessage()
-
+            print('I want.')
     def checkItemStatus(self):
-        if os.path.exists("predicted_class.txt") and self.weight_sensor.put_item:
-            self.object_classifier.pause_scanning()
-            if self.weight_sensor.is_item_added():
-                with open("predicted_class.txt", "r") as file:
-                    predicted_class_lines = file.readlines()
-                for predicted_class in predicted_class_lines:
-                    predicted_class = predicted_class.strip()
-                    if predicted_class:
-                        self.processScannedBarcode(predicted_class)
-                        print(predicted_class)
-                with open("predicted_class.txt", "w") as file:
-                    file.write('')
-                os.remove("predicted_class.txt")
-                self.weight_sensor.put_item = False
-                self.weight_sensor.verify = False
-                self.resumeScanningMessage()
-            
-        elif not self.weight_sensor.verify and self.weight_sensor.same_weight:
+        if self.weight_sensor.verify and not self.weight_sensor.same_weight:
             print('Error')
             message_box = QMessageBox()
             message_box.setWindowTitle("Shopping Paused!")
-            message_box.setText("Scan it to remove the item or put it back in the cart to continue.")
+            message_box.setText("Remove the added item to continue shopping!.")
             message_box.setStandardButtons(QMessageBox.Ok)
             message_box.exec_()
-            if os.path.exists("predicted_class.txt") and self.weight_sensor.remove_item:
+            self.weight_sensor.same_weight = False
+        else:
+            if os.path.exists("predicted_class.txt") and self.weight_sensor.put_item:
                 self.object_classifier.pause_scanning()
-                with open("predicteds_class.txt", "r") as file:
-                    predicted_class_lines = file.readlines()
-                for predicted_class in predicted_class_lines:
-                    predicted_class = predicted_class.strip()
+                if self.weight_sensor.is_item_added():
+                    with open("predicted_class.txt", "r") as file:
+                        predicted_class = file.read().strip()
                     if predicted_class:
-                        print('received', predicted_class)
-                        reference_number = config.transaction_info.get('reference_number')
-                        self.removeProductRFID(reference_number, predicted_class)
-                        with open("predicteds_class.txt", "w") as file:
+                        self.processScannedBarcode(predicted_class)
+                        with open("predicted_class.txt", "w") as file:
                             file.write('')
-                        os.remove("predicteds_class.txt")
+                        os.remove("predicted_class.txt")
+                        self.weight_sensor.put_item = False
+                        self.weight_sensor.verify = False
+                        self.resumeScanningMessage()
+            elif not self.weight_sensor.verify and self.weight_sensor.same_weight:
+                print('Error')
+                message_box = QMessageBox()
+                message_box.setWindowTitle("Shopping Paused!")
+                message_box.setText("Scan it to remove the item or put it back in the cart to continue.")
+                message_box.setStandardButtons(QMessageBox.Ok)
+                message_box.exec_()
+                if os.path.exists("predicted_class.txt") and self.weight_sensor.remove_item:
+                    self.object_classifier.pause_scanning()
+                    print('removed: classifier paused - 2')
+                    with open("predicted_class.txt", "r") as file:
+                        predicted_class = file.read().strip()
+                    if predicted_class:
+                        barcode = predicted_class
+                        print('received', barcode)
+                        reference_number = config.transaction_info.get('reference_number')
+                        self.removeProduct(reference_number, barcode)
+                        with open("predicted_class.txt", "w") as file:
+                            file.write('')
+                        os.remove("predicted_class.txt")
                         self.weight_sensor.same_weight = False
                         self.resumeScanningMessage()
 
@@ -451,13 +454,13 @@ class Ui_MainWindowItemView(object):
                                         translations[Config.current_language]['Scan_Barcode_Message'])
 
                 self.scanning_in_progress = True
-                last_scan_time = time.time()  
+                last_scan_time = time.time()  # Initialize last scan time
                 while self.scanning_in_progress:
                     ret, frame = self.cap.read()
                     if ret:
                         barcodes = decode(frame)
                         if barcodes:
-                            last_scan_time = time.time()
+                            last_scan_time = time.time()  # Update last scan time when a barcode is detected
                             for barcode in barcodes:
                                 barcode_data = barcode.data.decode("utf-8")
                                 self.scan_sound.play()
@@ -538,48 +541,6 @@ class Ui_MainWindowItemView(object):
             error_message = f"An unexpected error occurred while processing scanned barcode: {e}"
             print(error_message)
             QtWidgets.QMessageBox.critical(None, "Error", error_message)
-    
-    def processScannedRFID(self, predicted_class):
-        try:
-            product_details = self.db_manager.get_product_details_by_RFID(self.cursor, predicted_class)
-            if product_details:
-                product_id = product_details[0]
-                product_name = product_details[1]
-                product_price = product_details[-1]
-
-                identifier = f"{config.transaction_info.get('reference_number')} - {self.transaction_counter}"
-                sales_trans = config.transaction_info.get('reference_number')
-
-                existing_product = self.db_manager.checkProductInShoppingList(product_id)
-                if existing_product:
-                    self.db_manager.updateCartQuantity(product_id)
-                    print(f"CartQuantity updated for ProductId {product_id}.")
-                else:
-                    print(f"Product with ProductId {product_id} not found in the ShoppingListDetail table.")
-
-                self.productTable.insertRow(0)
-
-                item_name = QtWidgets.QTableWidgetItem(product_name)
-                item_price = QtWidgets.QTableWidgetItem(f"¥ {product_price:.2f}")
-                item_barcode = QtWidgets.QTableWidgetItem(str(predicted_class))
-                item_transaction = QtWidgets.QTableWidgetItem(str(identifier))
-                transaction_text = item_transaction.text()
-                self.db_manager.saveTransactionDetail(product_name, product_price, predicted_class, sales_trans,
-                                                      transaction_text)
-
-                self.productTable.setItem(0, 0, item_name)
-                self.productTable.setItem(0, 1, item_price)
-                
-                self.productTable.setItem(0, 2, item_barcode)
-
-                self.updateSummaryLabels()
-                print(f"Product with barcode {predicted_class} found in the database.")
-            else:
-                print(f"Product with barcode {predicted_class} not found in the database.")
-        except Exception as e:
-            error_message = f"An unexpected error occurred while processing scanned barcode: {e}"
-            print(error_message)
-            QtWidgets.QMessageBox.critical(None, "Error", error_message)
 
     def removeProduct(self, reference_number, barcode):
         try:
@@ -591,23 +552,6 @@ class Ui_MainWindowItemView(object):
                 if item and item.text() == barcode:
                     self.productTable.removeRow(row)
                     print(f"Row corresponding to Barcode {barcode} removed from the table.")
-                    self.updateSummaryLabels()
-                    break 
-        except Exception as e:
-            error_message = f"An unexpected error occurred while removing product: {e}"
-            print(error_message)
-            QtWidgets.QMessageBox.critical(None, "Error", error_message)
-    
-    def removeProductRFID(self, reference_number, predicted_class):
-        try:
-            self.db_manager.deleteTransactionDetailRFID(reference_number, predicted_class)
-            print(f"Product with ReferenceNumber {reference_number} and Barcode {predicted_class} removed.")
-
-            for row in range(self.productTable.rowCount()):
-                item = self.productTable.item(row, 2) 
-                if item and item.text() == predicted_class:
-                    self.productTable.removeRow(row)
-                    print(f"Row corresponding to Barcode {predicted_class} removed from the table.")
                     self.updateSummaryLabels()
                     break 
         except Exception as e:
